requirements.txt
# Core chess library
chess==1.11.2

# SVG to PNG conversion
cairosvg==2.7.1

# Zstandard decompression (for Lichess DB)
zstandard==0.22.0

# Environment variables
python-dotenv==1.0.0

# Image handling
pillow==10.1.0

# Progress bar for imports
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.1
mypy==1.7.1
